+++ app/quickfiles/route.ts
@@ -3,7 +3,7 @@ import { inject as service } from '@ember/service';
 import CasAuthenticatedRouteMixin from 'ember-osf-web/mixins/cas-authenticated-route';
 
 export default class Quickfiles extends Route.extend(CasAuthenticatedRouteMixin) {
-    currentUser = service('currentUser');
+    currentUser = service('current-user');
 
     model(this: Quickfiles) {
         return this.get('currentUser').get('currentUserId');